逻辑删除功能下放到 core 包,默认集成,只需要注解即可开启
