+++ mybatis-plus-extension/src/main/java/com/baomidou/mybatisplus/extension/injector/LogicSqlInjector.java
@@ -15,30 +15,7 @@
  */
 package com.baomidou.mybatisplus.extension.injector;
 
-import static java.util.stream.Collectors.toList;
-
-import java.util.List;
-import java.util.stream.Stream;
-
-import com.baomidou.mybatisplus.core.injector.AbstractMethod;
-import com.baomidou.mybatisplus.core.injector.AbstractSqlInjector;
-import com.baomidou.mybatisplus.core.injector.methods.Insert;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicDelete;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicDeleteBatchByIds;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicDeleteById;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicDeleteByMap;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectBatchByIds;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectById;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectByMap;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectCount;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectList;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectMaps;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectMapsPage;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectObjs;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectOne;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicSelectPage;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicUpdate;
-import com.baomidou.mybatisplus.extension.injector.methods.LogicUpdateById;
+import com.baomidou.mybatisplus.core.injector.DefaultSqlInjector;
 
 /**
  * SQL 逻辑删除注入器
@@ -46,28 +23,6 @@ import com.baomidou.mybatisplus.extension.injector.methods.LogicUpdateById;
  * @author hubin
  * @since 2018-06-12
  */
-public class LogicSqlInjector extends AbstractSqlInjector {
-
-    @Override
-    public List<AbstractMethod> getMethodList(Class<?> mapperClass) {
-        return Stream.of(
-            new Insert(),
-            new LogicDelete(),
-            new LogicDeleteByMap(),
-            new LogicDeleteById(),
-            new LogicDeleteBatchByIds(),
-            new LogicUpdate(),
-            new LogicUpdateById(),
-            new LogicSelectById(),
-            new LogicSelectBatchByIds(),
-            new LogicSelectByMap(),
-            new LogicSelectOne(),
-            new LogicSelectCount(),
-            new LogicSelectMaps(),
-            new LogicSelectMapsPage(),
-            new LogicSelectObjs(),
-            new LogicSelectList(),
-            new LogicSelectPage()
-        ).collect(toList());
-    }
+@Deprecated
+public class LogicSqlInjector extends DefaultSqlInjector {
 }
