--- mybatis-plus-extension/src/main/java/com/baomidou/mybatisplus/extension/injector/methods/LogicDeleteByIdWithFill.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2011-2019, hubin (jobob@qq.com).
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.baomidou.mybatisplus.extension.injector.methods;
-
-/**
- * use {@link com.baomidou.mybatisplus.extension.injector.methods.additional.LogicDeleteByIdWithFill}
- *
- * @author miemie
- * @since 2018-09-13
- * @deprecated 2018-11-09
- */
-@Deprecated
-public class LogicDeleteByIdWithFill extends com.baomidou.mybatisplus.extension.injector.methods.additional.LogicDeleteByIdWithFill {
-}
